% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyLP.R
\docType{package}
\name{FuzzyLP}
\alias{FuzzyLP}
\alias{FuzzyLP-package}
\title{Fuzzy Linear Programming}
\description{
FuzzyLP is a package to solve fuzzy linear programming problems.
}
\details{
FuzzyLP implements several algorithms for solving fuzzy linear programming
}
\references{
Villacorta, P.J., Rabelo, C.A., Pelta, D.A., and Verdegay, J.L. FuzzyLP: An R Package for Solving Fuzzy Linear Programming Problems. In: Kacprzyk J., Filev D., Beliakov G. (eds) Granular, Soft and Fuzzy Approaches for Intelligent Systems:209-230. Springer, 2017.

Cadenas, J.M. and Verdegay, J.L. PROBO: an interactive system in fuzzy linear programming. Fuzzy Sets and Systems 76(3):319-332, 1995.

Cadenas, J.M. and Verdegay, J.L. Modelos de Optimizacion con Datos Imprecisos. Servicio de Publicaciones, Universidad de Murcia, 1999.

Bellman, R. and Zadeh, L. Decision making in a fuzzy environment. Management Science 17 (B) 4:141-164, 1970.
}
