% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{FuzzyLP}
\alias{FuzzyLP}
\alias{FuzzyLP-package}
\title{Fuzzy Linear Programming}
\description{
FuzzyLP is a package to solve fuzzy linear programming problems.
}
\details{
FuzzyLP implements several algorithm for solving fuzzy linear programming,
the algorithms are fundamentally based in the book of...
}
\references{
Cadenas, J.M. and Verdegay, J.L. PROBO: an interactive system in fuzzy linear programming. Fuzzy Sets and Systems 76(3):319-332, 1995.

Cadenas, J.M. and Verdegay, J.L. Modelos de Optimizacion con Datos Imprecisos. Servicio de Publicaciones, Universidad de Murcia, 1999.

Bellman, R. and Zadeh, L. Decision making in a fuzzy environment. Management Science 17 (B) 4, pages 141-164, 1970.
}

